+++ plugin/src/main/kotlin/ru/cian/rustore/publish/service/RustoreBuildFormat.kt
@@ -1,0 +1,6 @@
+package ru.cian.rustore.publish.service
+
+enum class RustoreBuildFormat(val fileExtension: String) {
+    APK("apk"),
+    AAB("aab"),
+}