+++ plugin/src/main/kotlin/ru/cian/rustore/publish/service/RustoreServiceImpl.kt
@@ -14,15 +14,15 @@ import ru.cian.rustore.publish.models.response.UploadAppFileResponse
 import ru.cian.rustore.publish.utils.Logger
 import java.io.File
 
-private const val DOMAIN_URL = "https://public-api.rustore.ru"
-
 @SuppressWarnings("StringLiteralDuplication", "TooManyFunctions")
-internal class RustoreServiceImpl constructor(
-    private val logger: Logger
+internal class RustoreServiceImpl(
+    private val logger: Logger,
+    private val baseEntryPoint: String,
+    private val requestTimeout: Long?,
 ) : RustoreService {
 
     private val gson = Gson()
-    private val httpClient = HttpClientHelper(logger)
+    private val httpClient = HttpClientHelper(logger, requestTimeout)
 
     override fun getToken(
         keyId: String,
@@ -38,7 +38,7 @@ internal class RustoreServiceImpl constructor(
 
         logger.i("""
             curl --location --request POST \
-            $DOMAIN_URL/public/auth/ \
+            $baseEntryPoint/public/auth/ \
             --header 'Content-Type: application/json' \
             --data-raw '{
                 "keyId": "$keyId",
@@ -48,7 +48,7 @@ internal class RustoreServiceImpl constructor(
         """.trimIndent())
 
         val response = httpClient.post<AccessTokenResponse>(
-            url = "$DOMAIN_URL/public/auth/",
+            url = "$baseEntryPoint/public/auth/",
             body = gson.toJson(bodyRequest).toRequestBody(),
             headers = mapOf(
                 "Content-Type" to "application/json",
@@ -68,7 +68,7 @@ internal class RustoreServiceImpl constructor(
 
         logger.i("""
             curl --location --request POST \
-            $DOMAIN_URL/public/v1/application/$applicationId/version \
+            $baseEntryPoint/public/v1/application/$applicationId/version \
             --header 'Content-Type: application/json' \
             --header 'Public-Token: $token' \
             --data-raw '{
@@ -77,7 +77,7 @@ internal class RustoreServiceImpl constructor(
         """.trimIndent())
 
         val response = httpClient.post<AppDraftResponse>(
-            url = "$DOMAIN_URL/public/v1/application/$applicationId/version",
+            url = "$baseEntryPoint/public/v1/application/$applicationId/version",
             body = gson.toJson(bodyRequest).toRequestBody(),
             headers = mapOf(
                 "Content-Type" to "application/json",
@@ -118,39 +118,64 @@ internal class RustoreServiceImpl constructor(
         return response.body
     }
 
-    override fun uploadBuildFile(
+    override fun uploadApkBuildFile(
         token: String,
         applicationId: String,
         mobileServicesType: String,
         versionId: Int,
-        buildFile: File
+        artifactFormat: RustoreBuildFormat,
+        buildFile: File,
     ) {
 
         val fileBody = buildFile.asRequestBody(HttpClientHelper.MEDIA_TYPE_AAB)
-        val requestBody = MultipartBody.Builder()
+        val multipartBuilder = MultipartBody.Builder()
             .setType(MultipartBody.FORM)
             .addFormDataPart("file", buildFile.name, fileBody)
-            .addFormDataPart("servicesType", mobileServicesType)
-            .addFormDataPart("isMainApk", "true")
-            .build()
+
+        val requestFormArgument = artifactFormat.fileExtension
 
         val headers = mutableMapOf(
             "accept" to "application/json",
             "Public-Token" to token,
         )
 
-        logger.i("""
-            curl --location --request POST \
-            --header 'Content-Type: application/json' \
-            --header 'Public-Token: $token' \            
-            --form servicesType=$mobileServicesType \
-            --form isMainApk=true \
-            --form file='@${buildFile.absolutePath}' \
-            $DOMAIN_URL/public/v1/application/$applicationId/version/$versionId/apk
-        """.trimIndent())
+        when (artifactFormat) {
+            RustoreBuildFormat.APK -> {
+
+                multipartBuilder
+                    .addFormDataPart("servicesType", mobileServicesType)
+                    .addFormDataPart("isMainApk", "true")
+
+                logger.i(
+                    """
+                    curl --location --request POST \
+                    --header 'Content-Type: application/json' \
+                    --header 'Public-Token: $token' \            
+                    --form servicesType=$mobileServicesType \
+                    --form isMainApk=true \
+                    --form file='@${buildFile.absolutePath}' \
+                    $baseEntryPoint/public/v1/application/$applicationId/version/$versionId/$requestFormArgument
+                    """.trimIndent()
+                )
+            }
+            RustoreBuildFormat.AAB -> {
+                logger.i(
+                    """
+                    curl --location --request POST \
+                    --header 'Content-Type: application/json' \
+                    --header 'Public-Token: $token' \            
+                    --form file='@${buildFile.absolutePath}' \
+                    $baseEntryPoint/public/v1/application/$applicationId/version/$versionId/$requestFormArgument
+                    """.trimIndent()
+                )
+            }
+        }
+
+        val requestBody = multipartBuilder
+            .build()
 
         val response = httpClient.post<UploadAppFileResponse>(
-            url = "$DOMAIN_URL/public/v1/application/$applicationId/version/$versionId/apk",
+            url = "$baseEntryPoint/public/v1/application/$applicationId/version/$versionId/$requestFormArgument",
             body = requestBody,
             headers = headers
         )
@@ -164,6 +189,7 @@ internal class RustoreServiceImpl constructor(
         }
     }
 
+    @SuppressWarnings("MaxLineLength")
     override fun submit(
         token: String,
         applicationId: String,
@@ -172,13 +198,12 @@ internal class RustoreServiceImpl constructor(
     ): Boolean {
         logger.i("""
             curl --location --request POST \
-            $DOMAIN_URL/public/v1/application/$applicationId/version/$versionId/commit?priorityUpdate=$priorityUpdate \
+            $baseEntryPoint/public/v1/application/$applicationId/version/$versionId/commit?priorityUpdate=$priorityUpdate \
             --header 'Content-Type: application/json'
         """.trimIndent())
 
         val response = httpClient.post<SubmitPublicationResponse>(
-            url = "$DOMAIN_URL/public/v1/application/$applicationId/version/$versionId/commit" +
-                "?priorityUpdate=$priorityUpdate",
+            url = "$baseEntryPoint/public/v1/application/$applicationId/version/$versionId/commit?priorityUpdate=$priorityUpdate",
             body = "".toRequestBody(),
             headers = mapOf(
                 "Content-Type" to "application/json",
@@ -195,13 +220,13 @@ internal class RustoreServiceImpl constructor(
     ): Boolean {
 
         logger.i("""
-            curl --location --request DELETE $DOMAIN_URL/public/v1/application/$packageName/version/$previousAppId \
+            curl --location --request DELETE $baseEntryPoint/public/v1/application/$packageName/version/$previousAppId \
             --header 'Content-Type: application/json' \
             --header 'Public-Token: $token'            
         """.trimIndent())
 
         val response = httpClient.delete<DeleteAppDraftResponse>(
-            url = "$DOMAIN_URL/public/v1/application/$packageName/version/$previousAppId",
+            url = "$baseEntryPoint/public/v1/application/$packageName/version/$previousAppId",
             body = "".toRequestBody(),
             headers = mapOf(
                 "Content-Type" to "application/json",
@@ -212,5 +237,9 @@ internal class RustoreServiceImpl constructor(
         logger.v("response=$response")
 
         return response.code == "OK"
+    }
+
+    companion object {
+        const val DOMAIN_URL = "https://public-api.rustore.ru"
     }
 }