+++ plugin/src/main/kotlin/ru/cian/rustore/publish/utils/Consts.kt
@@ -1,3 +1,4 @@
 package ru.cian.rustore.publish.utils
 
 const val RELEASE_DATE_TIME_FORMAT = "yyyy-MM-dd'T'HH:mm:ssZZ"
+const val DATETIME_FORMAT_ISO8601 = "yyyy-MM-dd'T'HH:mm:ss.SSSSSSSXXX"